+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Bug fixes
+
+-   We now ignore errors thrown while sending verification emails on mount as intended
+
 ## [0.25.0]
+
 ### Additions
 
 -   Added `getRedirectURL` configuration option to third-party providers. This can be used to set where the user is redirected back during the callback.